fix chanId

+++ crypto/src/main/java/org/web3j/crypto/SignedRawTransaction.java
@@ -29,7 +29,7 @@ public class SignedRawTransaction extends RawTransaction {
         if (null == chainId) {
             encodedTransaction = TransactionEncoder.encode(this);
         } else {
-            encodedTransaction = TransactionEncoder.encode(this, chainId.byteValue());
+            encodedTransaction = TransactionEncoder.encode(this, chainId);
         }
         BigInteger v = Numeric.toBigInt(getSignatureData().getV());
         byte[] r = signatureData.getR();
